Relocate tests and use fixtures
